+++ roles/web-app-jitsi/files/playwright/test-guest-persona.js
@@ -1,0 +1,7 @@
+const { test } = require("@playwright/test");
+
+exports.register = function (shared) {
+  test("guest: public-landing → auth chain → never authenticated", async ({ page }) => {
+    await shared.runGuestFlow(page);
+  });
+};
