+++ roles/web-app-jitsi/files/playwright/test-login-biber.js
@@ -1,0 +1,12 @@
+const { test } = require("@playwright/test");
+
+exports.register = function (shared) {
+  test("biber: app → keycloak → join Jitsi room → logout", async ({ page }) => {
+    test.setTimeout(180_000);
+    await shared.runBiberFlow(page, {
+      biberInteraction: async (p) => {
+        await shared.reachJitsiPrejoin(p, "biber", "biber-room");
+      },
+    });
+  });
+};
